refactor: move ValidationResult into application/validation.py
